+++ scripts/console-helper.js
@@ -1,10 +1,10 @@
-const [owner, user] = await ethers
-  .getSigners()
-  .then(([owner, user]) =>
-    Promise.all([owner.getAddress(), user.getAddress()])
-  );
+const [_owner, userSigner] = await ethers.getSigners();
 
-const factory = await ethers.getContractFactory("Testament");
+const [owner, user, wife, son, attorney] = await ethers
+  .getSigners()
+  .then((signers) => Promise.all(signers.map((signer) => signer.getAddress())));
+
+const factory = await ethers.getContractFactory("Testament", userSigner);
 const testament = await factory.attach(contractAddress);
 
 const GOLD = await testament.GOLD();